// src/database/database.interfaces.ts
import { Config as IOptionsArangoDB } from 'arangojs/connection';
import {
  Abstract,
  DynamicModule,
  ForwardReference,
  Provider,
} from '@nestjs/common';

export interface ModuleMetadata {
  /**
   * Optional list of imported modules that export the providers which are
   * required in this module.
   */
  imports?: Array<
    Type<any> | DynamicModule | Promise<DynamicModule> | ForwardReference
  >;
  /**
   * Optional list of controllers defined in this module which have to be
   * instantiated.
   */
  controllers?: Type<any>[];
  /**
   * Optional list of providers that will be instantiated by the Nest injector
   * and that may be shared at least across this module.
   */
  providers?: Provider[];
  /**
   * Optional list of the subset of providers that are provided by this module
   * and should be available in other modules which import this module.
   */
  exports?: Array<
    | DynamicModule
    | Promise<DynamicModule>
    | string
    | symbol
    | Provider
    | ForwardReference
    | Abstract<any>
    | Function
  >;
}

export interface Type<T = any> extends Function {
  new (...args: any[]): T;
}

export interface IOptionsDatabase {
  arangodb: IOptionsArangoDB;
}

export interface DatabaseOptionsFactory {
  createDatabaseOptions(): Promise<IOptionsDatabase> | IOptionsDatabase;
}

export interface IOptionsDatabaseAsync extends Pick<ModuleMetadata, 'imports'> {
  useExisting?: Type<DatabaseOptionsFactory>;
  useClass?: Type<DatabaseOptionsFactory>;
  useFactory?: (...args: any[]) => Promise<IOptionsDatabase> | IOptionsDatabase;
  inject?: any[];
}